requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
attrs==21.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
idna
iniconfig==1.1.1
lxml==4.6.3
packaging==21.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
python-decouple==3.4
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
toml==0.10.2
urllib3==1.26.6
